requirements.txt
aiohttp==3.7.4.post0
aiohttp-cors==0.7.0
aioredis==1.3.1
ansible==4.0.0
ansible-core==2.11.0
appdirs==1.4.3
asgiref==3.3.4
astroid==2.4.2
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.9.3
blessings==1.7
boto3==1.17.82
botocore==1.20.82
branca==0.4.2
bs4==0.0.1
CacheControl==0.12.6
cachetools==4.2.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
click==7.1.2
colorama==0.4.3
colorful==0.5.4
contextlib2==0.6.0
coverage==5.1
cryptography==3.4.7
dictdiffer==0.8.1
distlib==0.3.0
distro==1.4.0
Django==3.2.4
django-admin-list-filter-dropdown==1.0.3
django-debug-toolbar==3.2.1
django-environ==0.4.5
django-filter==2.4.0
django-otp==1.0.2
django-query-profiler==0.8
django-storages==1.9.1
djangorestframework==3.12.4
docutils==0.15.2
entrypoints==0.3
et-xmlfile==1.0.1
filelock==3.0.12
flake8==3.7.9
folium==0.12.1
fsspec==2021.4.0
google-api-core==1.24.1
google-auth==1.24.0
googleapis-common-protos==1.52.0
gpustat==0.6.0
grpcio==1.34.0
hiredis==1.1.0
html5lib==1.0.1
idna==2.9
idna-ssl==1.1.0
ipaddr==2.2.0
isort==5.6.4
jdcal==1.4.1
Jinja2==2.11.3
jmespath==0.9.5
jsonschema==3.2.0
lazy-object-proxy==1.4.3
lockfile==0.12.2
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==1.1.1
mccabe==0.6.1
mmh3==3.0.0
mo-future==5.17.21182
modin==0.8.2
moz-sql-parser==3.32.20026
msgpack==1.0.2
multidict==4.7.5
mysqlclient==1.4.6
numpy==1.19.2
nvidia-ml-py3==7.352.0
opencensus==0.7.11
opencensus-context==0.1.2
openpyxl==3.0.3
packaging==20.3
pandas==1.1.4
pep517==0.8.2
progress==1.5
prometheus-client==0.9.0
protobuf==3.14.0
psutil==5.8.0
py-spy==0.3.3
pyarrow==1.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
pylint==2.6.0
pyparsing==2.3.1
pyrsistent==0.17.3
python-dateutil==2.8.1
pytoml==0.1.21
pytz==2019.3
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
resolvelib==0.5.4
retrying==1.3.3
rsa==4.7.2
s3transfer==0.4.2
sentry-sdk==0.14.2
six==1.14.0
soupsieve==2.0.1
sqlparse==0.3.0
toml==0.10.2
typing-extensions==3.7.4.1
urllib3==1.26.5
webencodings==0.5.1
wrapt==1.12.1
XlsxWriter==1.2.8
yarl==1.4.2
git+git://github.com/mplanes/django-authtools
